@@ -6,14 +6,19 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.View;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
+
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
 public class ConnectActivity extends AppCompatActivity {
 
@@ -26,18 +31,15 @@ public class ConnectActivity extends AppCompatActivity {
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
+
         sharedpreferences = getSharedPreferences(getString(R.string.MyPREFERENCES), Context.MODE_PRIVATE);
 
         final Button button = (Button) findViewById(R.id.connectButton);
         final EditText macAddText = (EditText) findViewById(R.id.editMacAddText);
         macAddText.setText(sharedpreferences.getString(getString(R.string.EV3KEY),""));
 
-
-
-
         button.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
-
 
                 Log.d("TAG", "onClick: "+ macAddText.getText().toString());
 
@@ -69,10 +71,7 @@ public class ConnectActivity extends AppCompatActivity {
             }
         });
     }
-
     private boolean validMacAdd(String macAdd) {
         return macAdd.length() == 17;
     }
-
-
 }
