@@ -6,19 +6,21 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.os.Parcelable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import java.io.Serializable;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
 import java.util.ArrayList;
 
 public class RCActivity extends AppCompatActivity {
@@ -26,15 +28,15 @@ public class RCActivity extends AppCompatActivity {
     BT_Comm btComm;
     SharedPreferences sharedPreferences;
 
-
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_rc);
         //Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         //setSupportActionBar(toolbar);
+
+        listenMessageFromRover();
+        sendMessageToRover("i_detect_you");
 
         btComm = new BT_Comm();
         sharedPreferences = getSharedPreferences(getString(R.string.MyPREFERENCES), Context.MODE_PRIVATE);
@@ -72,7 +74,6 @@ public class RCActivity extends AppCompatActivity {
 
         //sharedPreferences = getSharedPreferences(getString(R.string.MyPREFERENCES), Context.MODE_PRIVATE);
 
-
         if(!btComm.initBT()){
             // User did not enable Bluetooth
             btDisabledToast.show();
@@ -87,49 +88,23 @@ public class RCActivity extends AppCompatActivity {
             startActivity(intent);
         }
 
+        // En cas de clic sur le bouton d'ouverture partielle
         final Button buttonOuverturePartielle = (Button) findViewById(R.id.buttonOuverturePartielle);
         buttonOuverturePartielle.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
-              /*  try {
-                    btComm.writeMessage((byte) 1);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-                */
-
-                try {
-                    ArrayList<String> params = new ArrayList<>();
-                    params.add("ouvPartielle");
-                    MessageBluetooth msg = new MessageBluetooth("commande", params);
-                    btComm.writeMessage2(msg);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-
+                ouverturePartielle();
             }
         });
 
+        // En cas de clic sur le bouton d'ouverture totale
         final Button buttonOuvertureTotale = (Button) findViewById(R.id.buttonOuvertureTotale);
         buttonOuvertureTotale.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
-               /* try {
-                    btComm.writeMessage((byte) 2);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }*/
-
-                try {
-                    ArrayList<String> params = new ArrayList<>();
-                    params.add("ouvTotale");
-                    MessageBluetooth msg = new MessageBluetooth("commande", params);
-                    Log.d("MESSAGE OUV TOTALE", "onClick: "+msg.getParams());
-                    btComm.writeMessage2(msg);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
+                ouvertureTotale();
             }
         });
 
+        //Bouton permettant la déconnexion (pas encore implémenté)
         /*final Button buttonDeco = (Button) findViewById(R.id.button_deco);
         buttonDeco.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
@@ -138,6 +113,8 @@ public class RCActivity extends AppCompatActivity {
                     params.add("deconnexion");
                     MessageBluetooth msg = new MessageBluetooth("commande", params);
                     Log.d("MESSAGE DECONNEXION", "onClick: "+msg.getParams());
+
+                    //On envoie le message a la brique
                     btComm.writeMessage2(msg);
 
                     Intent intent = new Intent(RCActivity.this, ConnectActivity.class);
@@ -148,31 +125,37 @@ public class RCActivity extends AppCompatActivity {
             }
         });*/
 
-
         final EditText pseudo = (EditText) findViewById(R.id.textLogin);
         final EditText mdp = (EditText) findViewById(R.id.textPassword);
 
+        //Bouton permettant la création d'un utilisateur
         final Button creerUser = (Button) findViewById(R.id.createUserButton);
         creerUser.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
+                //On récupère les éléments du formulaire
                 String pseudotostring = pseudo.getText().toString();
                 String mdptostring = mdp.getText().toString();
-                User newuser = new User(pseudotostring,mdptostring);
-
+                User newuser = new User(pseudotostring, mdptostring);
 
                 try {
+                    //On créer un tableau de paramètres et on y rajoute les éléments
                     ArrayList<String> params = new ArrayList<>();
                     params.add(newuser.getName());
                     params.add(newuser.getPwd());
-                    MessageBluetooth msg = new MessageBluetooth("newuser", params);
-                    Log.d("MESSAGE", "onClick: "+msg.getParams());
-                    btComm.writeMessage2(msg);
-                    Log.d("Liste de user", ""+params);
 
+                    //On créer un nouveau message Bluetooth
+                    MessageBluetooth msg = new MessageBluetooth("newuser", params);
+                    Log.d("MESSAGE", "onClick: " + msg.getParams());
+
+                    //On envoie le message a la brique
+                    btComm.writeMessage2(msg);
+                    Log.d("Liste de user", "" + params);
+
+                    //On vide les champs du formulaire
                     pseudo.getText().clear();
                     mdp.getText().clear();
                     Toast.makeText(getApplicationContext(), "User créé", Toast.LENGTH_SHORT).show();
-                }catch (InterruptedException e) {
+                } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
@@ -184,30 +167,118 @@ public class RCActivity extends AppCompatActivity {
         final Button creerVehicule = (Button) findViewById(R.id.createVehiculeButton);
         creerVehicule.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
+                //On récupère les éléments du formulaire
                 String nameVehiToString = nameVehi.getText().toString();
                 String addMacToString = addMac.getText().toString();
-                Vehicule newvehicule = new Vehicule(nameVehiToString,addMacToString);
-
+                Vehicule newvehicule = new Vehicule(nameVehiToString, addMacToString);
 
                 try {
+                    //On créer un tableau de paramètres et on y rajoute les éléments
                     ArrayList<String> params = new ArrayList<>();
                     params.add(newvehicule.getNameVehi());
                     params.add(newvehicule.getAddMac());
-                    MessageBluetooth msg = new MessageBluetooth("newvehicule", params);
-                    Log.d("MESSAGE", "onClick: "+msg.getParams());
-                    btComm.writeMessage2(msg);
-                    Log.d("Liste de user", ""+params);
 
+                    //On créer un nouveau message Bluetooth
+                    MessageBluetooth msg = new MessageBluetooth("newvehicule", params);
+                    Log.d("MESSAGE", "onClick: " + msg.getParams());
+
+                    //On envoie le message a la brique
+                    btComm.writeMessage2(msg);
+                    Log.d("Liste de user", "" + params);
+
+                    //On vide les champs du formulaire
                     nameVehi.getText().clear();
                     addMac.getText().clear();
                     Toast.makeText(getApplicationContext(), "Véhicule créé", Toast.LENGTH_SHORT).show();
-                }catch (InterruptedException e) {
+                } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
         });
-
-
     }
 
+    //Gère l'envoi du message d'ouverture totale à la brique
+    public void ouvertureTotale() {
+        try {
+            ArrayList<String> params = new ArrayList<>();
+            params.add("ouvTotale");
+            MessageBluetooth msg = new MessageBluetooth("commande", params);
+            Log.d("MESSAGE OUV TOTALE", "onClick: "+msg.getParams());
+            btComm.writeMessage2(msg);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+
+    //Gère l'envoi du message d'ouverture partielle à la brique
+    public void ouverturePartielle() {
+        try {
+            ArrayList<String> params = new ArrayList<>();
+            params.add("ouvPartielle");
+            MessageBluetooth msg = new MessageBluetooth("commande", params);
+            btComm.writeMessage2(msg);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+
+    //Gère l'envoi du message de fermeture à la brique (pas implémenté)
+    /*public void fermeture_portail() {
+        try {
+            ArrayList<String> params = new ArrayList<>();
+            params.add("fermer"); // A Modifier
+            MessageBluetooth msg = new MessageBluetooth("commande", params);
+            btComm.writeMessage2(msg);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }*/
+
+    //Envoie un message sur firebase afin de prévenir le véhicule
+    public void sendMessageToRover(String msg) {
+        // Write a message to the database
+        FirebaseDatabase database = FirebaseDatabase.getInstance();
+        DatabaseReference myRef = database.getReference("portail");
+        myRef.setValue(msg);
+    }
+
+    //Ecoute firebase afin de s'avoir si le véhicule souhaite ouvrir ou fermer le portail
+    public void listenMessageFromRover() {
+        FirebaseDatabase database = FirebaseDatabase.getInstance();
+        DatabaseReference myRef = database.getReference("rover");
+        myRef.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                // This method is called once with the initial value and again
+                // whenever data at this location is updated.
+                String value = dataSnapshot.getValue(String.class);
+                Log.d("MESSAGE rover :", "Value is: " + value);
+                Context context = getApplicationContext();
+
+                if (!value.equals("null")) {
+                    Toast m = Toast.makeText(context, "MESSAGE rover : "+value, Toast.LENGTH_LONG);
+                    m.show();
+                }
+                sendMessageToRover("I receive : "+value);
+                //Action en fonction de ce que souhaite le véhicule
+                switch(value) {
+                    case "ouverture_total" :
+                        ouvertureTotale();
+                        break;
+                    case "ouverture_partielle" :
+                        ouverturePartielle();
+                        break;
+                    case "Fermeture_portail" :
+                        ouvertureTotale();
+                        break;
+                }
+            }
+
+            @Override
+            public void onCancelled(DatabaseError error) {
+                // Failed to read value
+                Log.w("MESSAGE rover", "Failed to read value.", error.toException());
+            }
+        });
+    }
 }
